Resolved some of the 2nd stage comment
