+++ apps/enterprise/views.py
@@ -11,7 +11,9 @@ from fyle_rest_auth.utils import AuthUtils
 from .models import Enterprise, Orgs, Exports
 from ..users.models import User
 from .serializers import ExportSerializer, EnterpriseSerializer, OrgsSerializer
-from ..exports import gsheet, fyle
+from ..exports import gsheet
+from ..fyle.utils import FyleConnector
+from fyle_finance_dashboard_api.utils import format_tpa, headers
 
 User = get_user_model()
 auth_utils = AuthUtils()
@@ -22,96 +24,9 @@ class EnterpriseView(viewsets.ViewSet):
     Fyle Finance Export
     """
 
-    permission_classes = [IsAuthenticated]
-
-    def post_export(self, request):
-
-        enterprise_id = request.data['id']
-        tpa_data_format = fyle.FormatData()
-        data_to_export = [tpa_data_format.headers()]
-        default_status = "Sync is in progress"
-        gsheet_object = gsheet.GoogleSpreadSheet()
-        sheet_id = None
-        try:
-            export = Exports.objects.get(enterprise_id=enterprise_id)
-            export.status = default_status
-            sheet_id = export.gsheet_link
-            export.save()
-        except Exception as e:
-            export = Exports(status=default_status, enterprise_id=enterprise_id)
-            export.save()
-            user = User.objects.get(user_id=request.user)
-            sheet_id = gsheet_object.create_sheet(user.email)
-
-        try:
-            orgs = Orgs.objects.filter(enterprise_id=enterprise_id)
-            for org in orgs.values():
-                fyle_tpa_data = fyle.FyleTpaData(org['refresh_token'])
-                data_to_export += tpa_data_format.format(fyle_tpa_data.fyle_tpa())
-            gsheet_object.write_data(data_to_export, sheet_id)
-            export = Exports.objects.get(enterprise_id=enterprise_id)
-            export.status = "Sync completed"
-            export.total_rows = len(data_to_export)-1
-            export.gsheet_link = sheet_id
-            export.save()
-            return Response(
-                data={
-                    'message': 'Sync completed',
-                    'data': len(orgs),
-                    'sheet_id': sheet_id,
-                },
-                status=status.HTTP_200_OK
-            )
-        except Exception as e:
-            return Response(
-                data={
-                    'message': 'No exports found'
-                },
-                status=status.HTTP_400_BAD_REQUEST
-            )
-
-    def get_orgs(self, request, **kwargs):
-        """
-        Get Orgs by enterprise id
-        """
-        orgs = Orgs.objects.filter(enterprise_id=kwargs['enterprise_id'])
-        if orgs:
-            return Response(
-                data=OrgsSerializer(orgs, many=True).data if orgs else [],
-                status=status.HTTP_200_OK
-            )
-        else:
-            return Response(
-                data={
-                    'message': 'Enterprise with this id does not exist'
-                },
-                status=status.HTTP_400_BAD_REQUEST
-            )
-
-    def get_export(self, request, **kwargs):
-        """
-        Get Exports by Id
-        """
-        try:
-            export = Exports.objects.get(enterprise_id=kwargs['enterprise_id'])
-            print (export)
-            return Response(
-                data=ExportSerializer(export).data,
-                status=status.HTTP_200_OK
-            )
-        except Exception as e:
-            print (e)
-            return Response(
-                data={
-                    'message': 'Exports not Available'
-                },
-                status=status.HTTP_400_BAD_REQUEST
-            )
-
     def post_enterprise(self, request, **kwargs):
 
         enterprise = Enterprise(name=request.data['name'])
-        print(enterprise)
         enterprise.save()
         if enterprise:
             return Response(
@@ -127,6 +42,9 @@ class EnterpriseView(viewsets.ViewSet):
             },
             status=status.HTTP_400_BAD_REQUEST
         )
+
+
+class OrgView(viewsets.ViewSet):
 
     def post_org(self, request, **kwargs):
 
@@ -155,6 +73,90 @@ class EnterpriseView(viewsets.ViewSet):
             },
             status= status.HTTP_400_BAD_REQUEST
         )
+
+    def get_orgs(self, request, **kwargs):
+        """
+        Get Orgs by enterprise id
+        """
+        orgs = Orgs.objects.filter(enterprise_id=kwargs['enterprise_id'])
+        if orgs:
+            return Response(
+                data=OrgsSerializer(orgs, many=True).data if orgs else [],
+                status=status.HTTP_200_OK
+            )
+        else:
+            return Response(
+                data={
+                    'message': 'Enterprise with this id does not exist'
+                },
+                status=status.HTTP_400_BAD_REQUEST
+            )
+
+
+class ExportView(viewsets.ViewSet):
+
+    def post_export(self, request):
+
+        enterprise_id = request.data['id']
+        data_to_export = [headers()]
+        default_status = "Sync is in progress"
+        gsheet_object = gsheet.GoogleSpreadSheet()
+        try:
+            export = Exports.objects.get(enterprise_id=enterprise_id)
+            export.status = default_status
+            sheet_id = export.gsheet_link
+            export.save()
+        except Exception as e:
+            export = Exports(status=default_status, enterprise_id=enterprise_id)
+            export.save()
+            user = User.objects.get(user_id=request.user)
+            sheet_id = gsheet_object.create_sheet(user.email)
+
+        try:
+            orgs = Orgs.objects.filter(enterprise_id=enterprise_id)
+            for org in orgs.values():
+                fyle_tpa_data = FyleConnector(org['refresh_token'])
+                data_to_export += format_tpa(fyle_tpa_data.get_fyle_tpa())
+            gsheet_object.write_data(data_to_export, sheet_id)
+            export = Exports.objects.get(enterprise_id=enterprise_id)
+            export.status = "Sync completed"
+            export.total_rows = len(data_to_export)-1
+            export.gsheet_link = sheet_id
+            export.save()
+            return Response(
+                data={
+                    'message': 'Sync completed',
+                    'data': len(orgs),
+                    'sheet_id': sheet_id,
+                },
+                status=status.HTTP_200_OK
+            )
+        except Exception as e:
+            print(e)
+            return Response(
+                data={
+                    'message': 'No exports found'
+                },
+                status=status.HTTP_400_BAD_REQUEST
+            )
+
+    def get_export(self, request, **kwargs):
+        """
+        Get Exports by Id
+        """
+        try:
+            export = Exports.objects.get(enterprise_id=kwargs['enterprise_id'])
+            return Response(
+                data=ExportSerializer(export).data,
+                status=status.HTTP_200_OK
+            )
+        except Exception as e:
+            return Response(
+                data={
+                    'message': 'Exports not Available'
+                },
+                status=status.HTTP_400_BAD_REQUEST
+            )
 
 
 class UserAccountMapping(generics.RetrieveAPIView):
