--- fyle-finance-dashboard-api/utils.py
@@ -1,15 +1,0 @@
-from rest_framework.views import Response
-from rest_framework.serializers import ValidationError
-
-
-def assert_valid(condition: bool, message: str) -> Response or None:
-    """
-    Assert conditions
-    :param condition: Boolean condition
-    :param message: Bad request message
-    :return: Response or None
-    """
-    if not condition:
-        raise ValidationError(detail={
-            'message': message
-        })
