+++ fyle_finance_dashboard_api/client_secret.json
@@ -1,0 +1,12 @@
+{
+  "type": "service_account",
+  "project_id": "multi-org-export",
+  "private_key_id": "6b3cc401116458996322bca843c33c293bb2f906",
+  "private_key": "-----BEGIN PRIVATE KEY-----\nMIIEvQIBADANBgkqhkiG9w0BAQEFAASCBKcwggSjAgEAAoIBAQCqxFhq+QCyD02m\nCmNie0sLmbghgcm5t1E4hYA7gJgXf8cgGQ37OtjjY3h79I9hKYZsOtlzGDStyDph\nOjkMPd5HKnGIFrAipfSKgtc9DnPUXVeSBwL7mu/UB20UgwsrRLPSSqP/9cshWuFd\ndFjBkKDNw1OTfA6Uw4MRVE/c29Yil16d03E7BkE23Knk3EeqCo3maweDJOWhWyJN\nX3wrI8ir9IzbiUEtWWgMwEhEPkJWhmCIbtfu7I4bfGuq+xbee4S5qYcfhwiZbHJm\nU3h28blmBMvuDebG1lM3CWbgC7+u49MxhUApj3XoL587LJwuUSEYtOM/N0Fzj/js\nO2EcN2b/AgMBAAECggEAC3/zdsJUAkNQXSK/9KBzulFDLbmtMwnquOqzBJuXvAvJ\nbIDwO+XuBMof3meAX4oWi/CFA+CY05PzUoIQPqHMxHAerljdl0e9XLW/XVSjoKTn\nwjHuSOSMy0BaX5I/mbQjtyd6cEDfyakMGjYJybQKqcpCNS40JF9SOUPudbCCqCUu\ne2DVPDmH0PjIGV6hvjgAjAKfaQek14FVrAqeaqFdEHSfrakgmNCmFlIsIznmOgYU\nKFcXrc753zFhHtjsTwiqAkdtUAEBz4Xs0chw+O4s4c3ZhtVYU7D2mzfI/t0guF8P\nD84JUth/9ryO1nGtOzFrQC75GEc6gWuKXTgZd4mKcQKBgQDR0bPrPxUBVKV0kuLg\nkKpB7ryaOgM1ALsQi4nV82QE4GTHpJmgd/VOy2Icy+47f9kEFGz7g91U8y2dUn2Y\niM5XdNt3m+5YHei66CtliAUvlJUF1BVtLuXTgC99cG0snfeeZjV688M6kKrK4ox7\ni6o4SP6rb+uV6O7HGr56JDGTVwKBgQDQWj2glPw1y+I/8xfoK1Z2CDnnqXOrmHWi\nVKPvfK7+76W2NaXjZQcrH2h9x4W2lkwEmPEqf7e8N1QFGdqMRR+VcBAm9Tu0zVcg\nL5RHhb2NrrUDR81YOhH8Q5hvUhloOrX4L8tHcz9BaDegR9gBWtykku6+y+CyqWeP\nxAzavT1omQKBgDAp9ycVP/kiOSjdXv81th+Uce7lSj8sf8R4g5d9W4RTDk9V1X4k\ni0cfINKDbZhy7JonJi1GQ9RwThRDD1mobdVdmdOHE9teYkIlcXDJHIejj+HaoWCC\n13cJgd1FpYoMaP8Pb09eDX6wHsSb5Kunj9eDyhIIiwfKhKVn2jmnfUnZAoGAG7qG\nzSJG1poFyGD/44QxA5BgtHYW0NV/glUlZDAB0ifus+s74qGFbLXHYEX3g69I+quo\nHPHWcBQk+HDSZyyj4W2CmMy2X3+cgsoSP9qKuZpwuLkLmRxRXHNajCYu/3Ig2aDy\nutWZq4jFLm2hT7zZ9IUhhP75Lo3hHH8I/uuEirECgYEArR4CdbMQ/o/VbORnqdAQ\nW9SWmhWDj9PqHnT58guD0V9p71BAypiGcjLx56DhtoVJw+V5qgrx/lcxT9Ar0atN\nqtkRaDx/Ul5QSwebZis/DGN15Xk84bydleICzN5QXO1mXgpgzk4CazEYAhkcAvQY\nvE1PCJVrPqLxTfk2vibBjWs=\n-----END PRIVATE KEY-----\n",
+  "client_email": "multiorgexport@multi-org-export.iam.gserviceaccount.com",
+  "client_id": "104273631591313885223",
+  "auth_uri": "https://accounts.google.com/o/oauth2/auth",
+  "token_uri": "https://oauth2.googleapis.com/token",
+  "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs",
+  "client_x509_cert_url": "https://www.googleapis.com/robot/v1/metadata/x509/multiorgexport%40multi-org-export.iam.gserviceaccount.com"
+}
