+++ fyle_finance_dashboard_api/urls.py
@@ -1,4 +1,4 @@
-"""fyle-finance-dashboard-api URL Configuration
+"""fyle_finance_dashboard_api URL Configuration
 
 The `urlpatterns` list routes URLs to views. For more information please see:
     https://docs.djangoproject.com/en/1.11/topics/http/urls/
