+++ fyle_finance_dashboard_api/wsgi.py
@@ -1,5 +1,5 @@
 """
-WSGI config for fyle-finance-dashboard-api project.
+WSGI config for fyle_finance_dashboard_api project.
 
 It exposes the WSGI callable as a module-level variable named ``application``.
 
@@ -11,6 +11,6 @@ import os
 
 from django.core.wsgi import get_wsgi_application
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "fyle-finance-dashboard-api.settings")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "fyle_finance_dashboard_api.settings")
 
 application = get_wsgi_application()
