+++ setup.sh
@@ -1,0 +1,18 @@
+#!/bin/bash
+# Django Settings
+export SECRET_KEY='#)^6o52k_j&wl$39k-e^w&@9%mi-o#c3*!p=+2cl(&4q@u384x'
+export ALLOWED_HOSTS='localhost'
+export DEBUG=True
+# Database Settings
+export DB_NAME='fyle_finance'
+export DB_USER='sinha'
+export DB_PASSWORD='password'
+export DB_HOST='localhost'
+export DB_PORT=''
+# Fyle Settings
+export API_URL='https://localhost:8000/api'
+export FYLE_BASE_URL='https://staging.fyle.in'
+export FYLE_CLIENT_ID='tpaPu6HJt8hfL'
+export FYLE_CLIENT_SECRET='1TzV6yhy6U'
+export FYLE_TOKEN_URI='https://accounts.staging.fyle.in/api/oauth/token'
+export FYLE_JOBS_URL='https://jobs.staging.fyle.in/v2/jobs'
